cmd/wdte: Fix API breakage and move importer() to its own file.
